Scrollbar: fix various issues
the rounding from 11b7dd8 was "rounding" an integer result instead of the float
some improved support for horizontal bars
scroll wheel support was wonky; essentially the reference point always being at the top of the nib made the scrolling buggy in one direction

issue #77

+++ gemrb/core/GUI/ScrollView.cpp
@@ -180,10 +180,10 @@ void ScrollView::ScrollbarValueChange(ScrollBar* sb)
 	const Point& origin = contentView.Origin();
 	
 	if (sb == hscroll) {
-		Point p(-sb->GetValue(), origin.y);
+		Point p(-short(sb->GetValue()), origin.y);
 		ScrollTo(p);
 	} else if (sb == vscroll) {
-		Point p(origin.x, -sb->GetValue());
+		Point p(origin.x, -short(sb->GetValue()));
 		ScrollTo(p);
 	} else {
 		Log(ERROR, "ScrollView", "ScrollbarValueChange for unknown scrollbar");
